✨ donation
Former-commit-id: 2f925fe62fa6151080b9321261126e44d4e2f6f1

+++ backend/src/main/java/com/yam/funteer/donation/request/DonationJoinReq.java
@@ -1,5 +1,12 @@
 package com.yam.funteer.donation.request;
 
+import java.time.LocalDateTime;
+
+import com.yam.funteer.pay.entity.Payment;
+import com.yam.funteer.pay.repository.PaymentRepository;
+import com.yam.funteer.post.entity.Post;
+import com.yam.funteer.user.entity.Member;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,5 +14,6 @@ import lombok.Setter;
 @Setter
 public class DonationJoinReq {
 	private Long memberId;
-	private Integer paymentAmount;
+	private Long paymentAmount;
+
 }
